third_party/blink/web_tests/platform/generic/http/tests/images/image-currentsrc-load-expected.txt

PASS img.currentSrc is ""
PASS img.width is 0
PASS img.currentSrc is "http://127.0.0.1:8000/resources/slow-image-dimensions.php?name=square100.png&mimeType=image/png&sleep=100"
PASS img.width is 100
PASS img.currentSrc is "http://127.0.0.1:8000/resources/slow-image-dimensions.php?name=square100.png&mimeType=image/png&sleep=100"
PASS img.width is 100
PASS img.currentSrc is "http://127.0.0.1:8000/resources/slow-image-dimensions.php?name=square200.png&mimeType=image/png&sleep=100"
PASS img.width is 200
PASS successfullyParsed is true

TEST COMPLETE
